% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_Mosa.R
\name{align_Mosa}
\alias{align_Mosa}
\title{Implement AROSICS in R}
\usage{
align_Mosa(
  path_in,
  ref_filepath,
  out_dir_path,
  corr_type = "global",
  max_shift = 250L,
  max_iter = 100L,
  grid_res = 1000L,
  window_size = NULL,
  window_pos = list(NULL, NULL),
  mp = NULL,
  save_data = TRUE,
  save_vector_plot = FALSE,
  apply_matrix = FALSE
)
}
\arguments{
\item{path_in}{chr. Path to the target image, or to a folder containing multiple target images. Images must be of Geotiff format.}

\item{ref_filepath}{chr. Path to the reference image.}

\item{out_dir_path}{chr. Directory where the outputs are saved.}

\item{corr_type}{chr. Type of co-registration. Either 'global' (default) or 'local'.}

\item{max_shift}{int. Maximum shift distance in reference image pixel units}

\item{max_iter}{int. Maximum number of iterations for matching (default: 5).}

\item{grid_res}{int. Tie point grid resolution in pixels of the target image (x-direction). Only applies to local co-registration.}

\item{window_size}{int. Custom matching window size (in pixels) as (X, Y) tuple (default: (1000, 1000)).}

\item{window_pos}{tuple. Custom matching window position as (X, Y) map coordinate in the same projection as the reference image (default: central position of image overlap). Only used when performing global co-registration.}

\item{mp}{int. Number of CPUs to use. If None (default), all available CPUs are used. If mp=1, no multiprocessing is done.}

\item{save_data}{logical. Saves the transformation metadata in a .pkl file, and the tie points data in a csv file. The latter only happens when performing local co-registration.}

\item{save_vector_plot}{logical. saves the a map of the calculated tie point grid in a JPEG file. Has an effect only when performing local co-registration.}

\item{apply_matrix}{logical. When correcting multiple images, applies the shifts computed for the first image to all the remaining ones, instead of computing the shifts independently. Allows for better alignement and faster computing time. WARNING : Currently, if inputs images don't all have the same extent, temporary padded images need to be created. We hope to change that soon ; until then, apply_matrix is disabled by default)}
}
\value{
None
}
\description{
Implementation in R of a pipeline that uses arosics(\href{https://github.com/GFZ/arosics}{AROSICS}) to perform a global or local co-registration on a file or a group of files located inside a folder
}
\examples{

\dontrun{
library(reticulate)

align_Mosa(path_in = "path_to_ortho.tif",
             ref_filepath = "ref_image.tif",
             out_dir_path = "my_output_dir",
             corr_type = "local",
             grid_res = 500,
             save_data = TRUE,
             save_vector_plot = TRUE
             )

align_Mosa(path_in = "path_to_input_folder",
             ref_filepath = "ref_image.tif",
             out_dir_path = "my_output_dir",
             corr_type = "local",
             grid_res = 500,
             save_data = FALSE,
             mp = 5,

             )

align_Mosa(path_in = "path_to_input_folder",
             ref_filepath = "ref_image.tif",
             out_dir_path = "my_output_dir",
             corr_type = "global",
             max_shift = 200,
             save_data = TRUE,
             apply_matrix = TRUE,
             mp = 1
             )
             }
}
