% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_Mosa.R
\name{apply_saved_matrix}
\alias{apply_saved_matrix}
\title{Applies an already saved AROSICS transform to a new image}
\usage{
apply_saved_matrix(im_path, out_dir_path, metadata_path, suffix = "_")
}
\arguments{
\item{im_path}{chr. Path to the target image, or to a folder containing multiple target images. Images must be of Geotiff format.}

\item{out_dir_path}{chr. Directory where the outputs are saved.}

\item{metadata_path}{chr. Path to the .pkl file containing info on the transform to apply. This file can be saved when applying the align_Mosa function with save_data set to TRUE.}

\item{suffix}{chr. Text to add at the end of the output filenames.}
}
\value{
None
}
\description{
Applies an already saved AROSICS transform to a new image
}
\examples{

\dontrun{
library(reticulate)

apply_saved_matrix(im_path = "path_to_ortho.tif",
                   out_dir_path = "my_output_dir",
                   metadata_path = "my_ortho_metadata.pkl",
                   )
                   }
}
