% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arosics_in_r.R
\name{arosics_in_r}
\alias{arosics_in_r}
\title{Implement AROSICS in R}
\usage{
arosics_in_r(
  path_in,
  ref_filepath,
  out_dir_path,
  corr_type = "global",
  max_shift = 250L,
  max_iter = 100L,
  grid_res = 1000L,
  window_size = NULL,
  window_pos = list(NULL, NULL),
  mp = NULL,
  save_data = TRUE,
  save_vector_plot = FALSE,
  dynamic_corr = FALSE,
  apply_matrix = FALSE
)
}
\arguments{
\item{path_in}{xx}

\item{ref_filepath}{xx}

\item{out_dir_path}{xx}

\item{corr_type}{xx}

\item{max_shift}{xx}

\item{max_iter}{xx}

\item{grid_res}{xx}

\item{window_size}{xx}

\item{window_pos}{xx}

\item{mp}{xx}

\item{save_data}{xx}

\item{save_vector_plot}{xx}

\item{dynamic_corr}{xx}

\item{apply_matrix}{xx}
}
\description{
R functions to implement the fonctions of the Python package \href{https://github.com/GFZ/arosics}{AROSICS}
}
