% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_availableCrowns.R
\name{check_availableCrowns}
\alias{check_availableCrowns}
\title{Check the number of crowns included in each image}
\usage{
check_availableCrowns(path_bbox, crownsFile, dates = NULL)
}
\arguments{
\item{path_bbox}{character vector. Path to the non NA Bbox return by the function \code{extract_bboxImages()}}

\item{crownsFile}{sf. Crowns polygons with an 'id' variable.}

\item{dates}{character vector. Dates (format '\%Y\%m\%d'), order matching \code{path_bbox}. If NULL, auto-generated.}
}
\value{
A named list of vectors of crown IDs included in each image.
}
\description{
As the extend of images can changed from one image to another, this function return
the number of crowns included in each image and the vector of the id contained for each id.
}
