% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_crownsFile.R
\name{check_crownFile}
\alias{check_crownFile}
\title{Check the crown file compatibility}
\usage{
check_crownFile(crownFile)
}
\arguments{
\item{crownFile}{A sf object}
}
\value{
Text that give you information about your file
}
\description{
Check the crown file compatibility
}
\examples{
library(sf)
library(dplyr)
mean_lat <- 46.07998
sd_lat <- 0.1
mean_long <- 8.931849
sd_long <-  0.1

set.seed(42)
dat_sim <- data.frame(lat = rnorm(3, mean = mean_lat, sd = sd_lat),
                      long = rnorm(3, mean = mean_long, sd = sd_long))

dat_sf <- sf::st_as_sf(dat_sim, coords = c("long", "lat"), crs = 4326) \%>\%
   sf::st_transform(3035)

# Buffer circles by 100m
crownFile <- sf::st_buffer(dat_sf, dist = 1000) \%>\%
   dplyr::mutate(id = c(122,202,122),
          family = c('Fabaceae', 'Ochnaceae', 'Fabaceae'),
          gen = c('Newtonia','Lophira','Guibourtia'),
          tx_sp_lvl = c('Newtonia leucocarpa','Lophira alata','Guibourtia tessmannii'),
          plot_name = 'mbalmayo_pheno_observatory',
          code_sp = c(12856, 1690, 5691))

check_crownFile(crownFile)
}
