% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{color_label}
\alias{color_label}
\title{Vector to associate color to label}
\format{
A vector
}
\usage{
data("color_label")
}
\description{
Vector to associate color to label in ggplot
}
\examples{
data('color_label')
}
\keyword{datasets}
\keyword{internal}
