% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_files_architecture.R
\name{create_files_architecture}
\alias{create_files_architecture}
\title{Create the recommended files architecture}
\usage{
create_files_architecture(main_folder_path = getwd(), main_folder_name = NULL)
}
\arguments{
\item{main_folder_path}{Character. The path where to create the main folder}

\item{main_folder_name}{Character. The name of the main folder. Must be provided.}
}
\description{
Create the recommended files architecture.
}
