% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_labelingFile.R
\name{create_labelingFile}
\alias{create_labelingFile}
\title{Create xlsx labeling file}
\usage{
create_labelingFile(crownsFile, site, dates, path_out = NULL)
}
\arguments{
\item{crownsFile}{sf. Crowns file}

\item{site}{chr. site name.}

\item{dates}{chr. vector of dates (format should be 'YYYY_MM_DD', p.e c('2022_09_25','2022_10_10').}

\item{directory}{The path where to save the xlsx file.}
}
\description{
A function to create the xlsx file to do the labeling
}
\examples{

library(canObsR)

data(crowns)

create_labelingFile(
crowns,
site = 'Site1',
dates = c('2025_05_10','2025_05_20'),
path_out = NULL
)

}
