% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nonNA_bbox.R
\name{create_nonNA_bbox}
\alias{create_nonNA_bbox}
\title{Extract and export the non NA area of images.}
\usage{
create_nonNA_bbox(path_in, dates = NULL, out_dir_path = NULL, filename = NULL)
}
\arguments{
\item{path_in}{character. Path to the main folder of the target images. Images must be of Geotiff format.}

\item{dates}{character vector. Dates (format of dates should be '\%Y-\%m-\%d', '\%Y\%m\%d' or '\%Y_\%m_\%d').The order should match \code{path_in}.}

\item{out_dir_path}{character. Directory where the outputs are saved.
One file will be export per date. If NULL, the vector will not be exported.}

\item{filename}{logical. Name of the files.}
}
\description{
This function extract the area of images covered by non NA values and return
an \code{sf} object with one polygon per image.
}
