% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nonNA_bboxImages.R
\name{create_nonNA_bboxImages}
\alias{create_nonNA_bboxImages}
\title{Extract and export the non NA area of images.}
\usage{
create_nonNA_bboxImages(
  path_in,
  dates = NULL,
  directory = NULL,
  filename = NULL
)
}
\arguments{
\item{path_in}{chr. Path to the folder where the mosaics are located.}

\item{dates}{chr. Vector with dates (format should be '\%Y\%m\%d', p.e
'20220925'). The order of the dates should match with the order of the
dates of the mosaics in the path_in folder.}

\item{directory}{chr. Path to the directory used to export the vector files.
One file will be export per date. If NULL, the vector will not be exported.}

\item{filename}{logical. Name of the files.}
}
\description{
This function extract the area of images covered by non NA values and return
an \code{sf} object with one polygon per image.
}
\examples{

\dontrun{

# my_directory <- "MY_PATH_FOR_OUTPUTS"

bbox <- create_nonNA_bboxImages(
path_in = system.file(package="managecrownsdata"), 'rgb/'),
directory = my_directory)

}

}
