% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{crowns}
\alias{crowns}
\title{crowns delineation data}
\format{
A sf object with 189 observations on the following 5 variables:
\itemize{
\item \code{id}: Crown id
\item \code{family}: family
\item \code{genus}: genus
\item \code{species}: species
\item \code{geom}: geometry
}
}
\usage{
data("crowns")
}
\description{
Crowns delineation data
}
\examples{
data(crowns)
}
\keyword{datasets}
\keyword{internal}
