% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{data}
\alias{data}
\title{Full crowns data}
\format{
A tibble with 131,656 observations on the following 23 variables:
\itemize{
\item \code{site} : site name
\item \code{id}: Crown id
\item \code{date}: date of the rgb survey
\item \code{family}: family
\item \code{genus}: genus
\item \code{species}: species
\item \code{type}: type
\item \code{metric}: metric
\item \code{band}: band
\item \code{value}: value7
\item \code{PPfoliar1} :
\item \code{PPfoliar2} :
\item \code{PPFlo} :
\item \code{PPFr} :
\item \code{PPFlo_uncertainty} :
\item \code{PPFr_uncertainty} :
\item \code{desynchr} :
\item \code{PPfoliar2_uncertainty} :
\item \code{obs}: Observer
\item \code{comments} : any comment
\item \code{update} : date of the last update
\item \code{Usable_crown} : Is the crowns considered as good ?
}
}
\usage{
data("data")
}
\description{
Full crowns data
}
\examples{
data('data')
}
\keyword{datasets}
\keyword{internal}
