% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{data_labeling}
\alias{data_labeling}
\title{Labeling data}
\format{
A data.frame with 189000 observations on the following 12 variables:
\itemize{
\item \code{site}: site
\item \code{id}: Crown id
\item \code{family}: family
\item \code{genus}: genus
\item \code{species}: species
\item \code{n}: Number of individuals per species
\item \code{obs}: Observer
\item \code{update}: date of the last update
\item \code{date}: date of the rgb survey
\item \code{phenophase}: labels
\item \code{comments}: comments
\item \code{Usable_crown}: Usable_crown
}
}
\usage{
data("data_labeling")
}
\description{
Labeling data
}
\examples{
data(data_labeling)
}
\keyword{datasets}
\keyword{internal}
