% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{extr_dates}
\alias{extr_dates}
\title{Extract dates from files names}
\usage{
extr_dates(names_img)
}
\arguments{
\item{names_img}{chr. The files basenames.}
}
\description{
Extract dates from files names
}
