% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_bboxImages.R
\name{extract_bboxImages}
\alias{extract_bboxImages}
\title{Extract and export the non NA area of images.}
\usage{
extract_bboxImages(path_in, dates = NULL, directory = NULL, filename = NULL)
}
\arguments{
\item{path_in}{vector with the full paths to the RGB rasters images.}

\item{dates}{chr. Vector with dates (format should be '\%Y\%m\%d', p.e
'20220925'). The order of the dates should match with the order of the
dates of the image in the path_in}

\item{directory}{chr. The path to the directory used to export the vectors.
One file will be export per date. If NULL, the vector will not be exported.}

\item{filename}{logical. The name of the files.}
}
\description{
This function extract the area of images covered by non NA values and return
an \code{sf} object with one polygon per image.
}
\examples{

rgb_paths <- list.files(
file.path(
system.file(package="managecrownsdata"), 'rgb/'),
full.names = TRUE
)

#bbox <- extract_bboxImages(
#path_in = rgb_paths,
#directory = my_directory)
}
