% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_RGB_values.R
\name{extract_RGB_values}
\alias{extract_RGB_values}
\title{Extract RGB metrics}
\usage{
extract_RGB_values(
  crowns,
  RGB_paths,
  site = NULL,
  date = NULL,
  fun = "all",
  infos = FALSE,
  crs = NULL,
  colname_id = "id",
  colname_family = "family",
  colname_genus = "genus",
  colname_specie = "specie",
  colname_plot_name = "plot_name",
  colname_code_sp = "code_sp"
)
}
\arguments{
\item{RGB_paths}{a list with the full paths to the RGB rasters.}

\item{site}{chr. name of the site, p.e 'Mbalmayo'.}

\item{date}{chr. vector of dates (format should be 'yyyy_mm_dd', p.e '2022_09_25').
The order of the dates should match with the order of the RGB_paths !}

\item{fun}{chr. Specify the function used in the 'fun' parameter of the \code{exactextractr::exact_extract} function to extract
RGB values, it could be 'mean', 'var' or 'all'.}

\item{infos}{logical. Specify whether or not to return details of the extraction.
When TRUE, specify the crowns which has not been extracting per date, because they were out of the image.}

\item{crs}{crs. Object of class 'crs', could be get from st_crs(..). If NULL, it will
use and transform all the data into the crs of the first RGB image.}

\item{colname_id}{chr. Specify the name of the variable id. By default 'id'}

\item{colname_family}{chr. Specify the name of the variable id. By default 'family'}

\item{colname_genus}{chr. Specify the name of the variable id. By default 'genus'}

\item{colname_specie}{chr. Specify the name of the variable id. By default 'specie'}

\item{colname_plot_name}{chr. Specify the name of the variable id. By default 'plot_name'}

\item{colname_code_sp}{chr. Specify the name of the variable id. By default 'code_sp'}
}
\description{
The function extracts RGB metrics (red, green, blue, sumrgb,
rcc, gcc, bcc, gndvi, gli) for each crown at each date. It first compute the metrics
for the whole mosaics and then extracts the value at the crown scale using the \code{exactextractr::exact_extract}
function. The mean and / or the variance can be extracted (see 'fun' parameter).
}
\examples{
\dontrun{
library(sf)
library(dplyr)
library(stringr)

path <- system.file(package="phenobsappli")
RGB_paths <- list.files(file.path(path, 'tif'), full.names = TRUE)
crowns <- read_sf (list.files(file.path(path, 'shp'), full.names = TRUE))
date = sapply( str_split( basename( RGB_paths ),'_' ),
                     function(x) str_remove(paste(x[2], x[3], x[4], sep ='_'),'.tif') )
crs = sf::st_crs(stars::read_stars(RGB_paths[1], proxy = TRUE))
site = 'Mbalmayo'

test <- extract_RGB_values(
   crowns,
   RGB_paths,
   crs = crs,
   site = site,
   date = date,
   fun = 'all',
   infos = TRUE,
   colname_id = 'id',
   colname_family = 'family',
   colname_genus = 'genus',
   colname_specie = 'specie',
   colname_plot_name = 'plot_name',
   colname_code_sp = 'code_sp'
)
}

}
