% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rgbValues.R
\name{extract_rgbValues}
\alias{extract_rgbValues}
\title{Extract spectral indices from RGB at the crown scale}
\usage{
extract_rgbValues(
  path_images,
  path_crowns,
  out_dir_path = NULL,
  N_cores = 1,
  sites = NULL,
  dates = NULL,
  tempdir_custom = NULL,
  file_type = ".RData"
)
}
\arguments{
\item{path_images}{character vector. Path to the target images. Images must be of Geotiff format.}

\item{path_crowns}{character. Path to the crowns polygons file.}

\item{out_dir_path}{character. Directory where the outputs are saved.}

\item{N_cores}{integer. Number of cores use in the parallelisation proccess.}

\item{sites}{character. Name of the site, p.e 'Mbalmayo'.}

\item{dates}{character vector. Dates (format of dates should be '\%Y-\%m-\%d', '\%Y\%m\%d' or '\%Y_\%m_\%d').The order should match \code{path_images}.}

\item{tempdir_custom}{character. Directory where the temporary files are saved.}

\item{file_type}{character. By default it is '.RData' but can be '.csv' or '.xlsx'}
}
\value{
A tibble with the variable site, id, date, family, genus, species, type, metric, band and value.
}
\description{
The function extracts RGB indices (red, green, blue, sumrgb,
rcc, gcc, bcc, gndvi, gli) for each crown at each date. It extracts the value
at the crown scale using the \code{exactextractr::exact_extract}
function. The mean and / or the variance can be extracted (see 'fun' parameter).
}
\examples{

\dontrun{

imgs = list.files('my-path-to-images', full.names = T)
path_crowns = "my-path-to-crowns-shapefile"
out_dir_path = "output-directory"

rgb_data <- extract_rgbValues (
  path_images = imgs,
  path_crowns = path_crowns,
  out_dir_path = out_dir_path,
  N_cores = 10,
  sites = NULL,
  dates = NULL
)
}

}
