% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_labels.R
\name{format_labels}
\alias{format_labels}
\title{Format labels values from wide data to long data}
\usage{
format_labels(wideLabels)
}
\arguments{
\item{wideLabels}{A \code{tibble} or \code{dataframe} which contains the labels directly import from the xlsx file.}
}
\value{
A \code{tibble}
}
\description{
A function to format labels data from wide to long format
}
