% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap_Labels.R
\name{Labels_heatmap}
\alias{Labels_heatmap}
\title{Plot the heatmap of Labels}
\usage{
Labels_heatmap(
  longLabels,
  Specie = NULL,
  Genus = NULL,
  Family = NULL,
  title = NULL,
  na.rm = FALSE
)
}
\arguments{
\item{longLabels}{A \code{tibble} or \code{dataframe} of long data format for labels}

\item{Specie}{chr}

\item{Genus}{chr}

\item{Family}{chr}

\item{title}{chr}

\item{na.rm}{logical Remove NA when TRUE}
}
\value{
return a ggplot
}
\description{
A fct function
}
