% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_Labels.R
\name{Labels_pivotLonger}
\alias{Labels_pivotLonger}
\title{Reformat table of labels from wide data to long data}
\usage{
Labels_pivotLonger(wideLabels)
}
\arguments{
\item{wideLabels}{A \code{tibble} or \code{dataframe} which contains the labels directly import from the xlsx file.}
}
\value{
A \code{tibble}
}
\description{
A function to format labels data from wide to long format
}
