% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_values.R
\name{merge_values}
\alias{merge_values}
\title{Merge labels and rgb values}
\usage{
merge_values(longLabels, rgbValues)
}
\arguments{
\item{longLabels}{\code{tibble} or \code{dataframe} which contains the labels directly import from the xlsx file.}

\item{rgbValues}{\code{tibble} or \code{dataframe} with the rgb metrics values.}
}
\value{
A \code{tibble}
}
\description{
A function to merge the labels from long format and the rgb metrics
values.
}
\examples{

library(sf)
library(dplyr)
library(readxl)

path_crownsFile <- file.path(
system.file(package="managecrownsdata"),
'crowns/Bouamir_crowns.gpkg')
crownsFile <- sf::read_sf(path_crownsFile)
rgb_paths <- list.files(
file.path(
system.file(package="managecrownsdata"), 'rgb/'),
full.names = TRUE
)

check_crownsFile(crownsFile = crownsFile)

crownsFile <- crownsFile \%>\% dplyr::rename(
   geometry = geom
)

check_crownsFile(crownsFile = crownsFile)

rgbValues <- extract_rgbValues(
crownsFile = crownsFile,
path_images = rgb_paths,
fun = 'all',
infos = FALSE
)

raw_labels <- read_excel(
file.path(
system.file(package="managecrownsdata"),
'xlsx/labeling_file_Bouamir.xlsx')
)
longLabels <- pivot_Labels(raw_labels, simplify_labels = TRUE)

merge_values(longLabels, rgbValues)

}
