% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_Labels.R
\name{pivot_Labels}
\alias{pivot_Labels}
\title{Format labels values from wide data to long data}
\usage{
pivot_Labels(wideLabels, simplify_labels = FALSE)
}
\arguments{
\item{crownsFile}{\code{sf object}. The crowns delineation file. It should include 'id',
'species', 'genus' and 'family' columns.}

\item{site}{chr. site name.}

\item{dates}{chr. vector of dates (format should be 'YYYY_MM_DD', p.e '20220925').}
}
\description{
A function to format labels data from wide to long format
}
\examples{

library(readxl)

raw_labels <- read_excel(file.path(system.file(package="managecrownsdata"), 'xlsx/labeling_file_Bouamir.xlsx'))

pivot_Labels(raw_labels)

pivot_Labels(raw_labels, simplify_labels = TRUE)

}
