% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmapLabels.R
\name{plot_heatmapLabels}
\alias{plot_heatmapLabels}
\title{Plot the heatmap of Labels}
\usage{
plot_heatmapLabels(
  longLabels,
  Specie = NULL,
  Genus = NULL,
  Family = NULL,
  title = NULL
)
}
\arguments{
\item{longLabels}{tibble}

\item{Specie}{chr}

\item{Genus}{chr}

\item{Family}{chr}

\item{title}{chr}
}
\value{
return a ggplot
}
\description{
A fct function
}
