% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signal.R
\name{plot_signal}
\alias{plot_signal}
\title{Plot the spectral signal over time}
\usage{
plot_signal(
  data,
  Species = NULL,
  Genus = NULL,
  Family = NULL,
  Type = NULL,
  Metric = "mean",
  Band = NULL,
  facet_by = "band",
  slcted_id = "all",
  title = NULL
)
}
\arguments{
\item{data}{A \code{tibble} or \code{dataframe} of long data format for labels}

\item{Species}{chr}

\item{Genus}{chr}

\item{Family}{chr}

\item{Type}{chr}

\item{Metric}{chr}

\item{Band}{chr}

\item{facet_by}{chr}

\item{slcted_id}{chr}

\item{title}{chr}
}
\value{
return a ggplot
}
\description{
A fct function
}
