% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmapLabels.R
\name{quali_heatmap}
\alias{quali_heatmap}
\title{Plot the heatmap of Labels}
\usage{
quali_heatmap(
  longLabels,
  Specie = NULL,
  Genus = NULL,
  Family = NULL,
  title = NULL
)
}
\arguments{
\item{longLabels}{tibble}

\item{specie}{chr}

\item{genus}{chr}

\item{family}{chr}
}
\value{
return a ggplot
}
\description{
A fct function
}
