% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_labels.R
\name{shiny_labels}
\alias{shiny_labels}
\title{Labeling shiny app}
\usage{
shiny_labels(data_labeling = NULL, newFile = NULL, imgFolder = NULL)
}
\arguments{
\item{data_labeling}{tbl_df. Labeling file resulting from the \code{\link[canObsR]{create_labelingFile}}}

\item{newFile}{character. Path to save the new labeling file.}

\item{imgFolder}{character. Path to crown images folder.}
}
\description{
Shiny app to do the labeling based on crowns images.
}
\examples{

library(canObsR)

data(data_labeling)
data_labeling
# shiny_labels(data_labeling = data_labeling) # Run it

}
