% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_visualisation.R
\name{shiny_visualisation}
\alias{shiny_visualisation}
\title{Visualisation shiny app}
\usage{
shiny_visualisation(data_labeling)
}
\arguments{
\item{data_labeling}{tibble. Labeling file resulting from the \code{\link[canObsR]{create_labelingFile}}
and modify from the \code{\link[canObsR]{shiny_labels}}}
}
\description{
Shiny app to do the visualize the crowns data (labels and rgb indices).
}
\examples{

library(canObsR)

# shiny_visualisation() # Run it

}
