% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time_SIFT_in_r.R
\name{Time_SIFT_in_r}
\alias{Time_SIFT_in_r}
\title{xxxx}
\usage{
Time_SIFT_in_r(
  pathDIR,
  out_dir_ortho,
  out_dir_DEM = NULL,
  out_dir_project = NULL,
  data_type = "RGB",
  resol_ref = 0.05,
  crs = "EPSG::32622",
  site_name = "",
  calibrate_col = TRUE,
  sun_sensor = FALSE,
  group_by_flight = FALSE,
  downscale_factor_alignement = 1L,
  downscale_factor_depth_map = 2L
)
}
\arguments{
\item{pathDIR}{xxx}

\item{out_dir_ortho}{xxx}

\item{out_dir_DEM}{xxx}

\item{out_dir_project}{xxx}

\item{data_type}{xxx}

\item{resol_ref}{xxx}

\item{crs}{xxx}

\item{site_name}{xxx}

\item{calibrate_col}{xxx}

\item{sun_sensor}{xxx}

\item{group_by_flight}{xxx}

\item{downscale_factor_alignement}{xxx}

\item{downscale_factor_depth_map}{xxx}
}
\description{
xxxx
}
