% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_availableCrowns.R
\name{check_availableCrowns}
\alias{check_availableCrowns}
\title{Check the number of crowns included in each image}
\usage{
check_availableCrowns(Bbox_path, crownFile, date = NULL, crs = NULL)
}
\arguments{
\item{crownFile}{A \code{sf} object for the crowns with an 'id' variable.}

\item{date}{chr. Vector with dates (format should be '\%Y_\%m_\%d', p.e
'2022_09_25'). The order of the dates should match with the order of the
dates of the image in the RGB_paths.}

\item{crs}{crs. Object of class 'crs', could be get from st_crs(..). If NULL,
it will use and transform all the data into the crs of the first non NA bbox.}

\item{path_in}{The path to the non NA Bbox return by the function \code{extract_bboxImages()}}
}
\value{
A list with all the crowns id contained within images. The names of the list elements
are the date of the images.
}
\description{
As the extend of images can changed from one image to another, this function return
the number of crowns included in each image and the vector of the id for crowns in images.
}
