% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_availableCrowns.R
\name{check_availableCrowns}
\alias{check_availableCrowns}
\title{Check the number of crowns included in each image}
\usage{
check_availableCrowns(path_bbox, crownFile, dates = NULL)
}
\arguments{
\item{path_bbox}{The path to the non NA Bbox return by the function \code{extract_bboxImages()}}

\item{crownFile}{\code{sf object}  for the crowns with an 'id' variable.}

\item{dates}{chr. Vector with dates (format should be '\%Y\%m\%d', p.e
'20220925'). The order of the dates should match with the order of the
dates of the image in the path_bbox}
}
\value{
\code{list} and \code{dotchart}
}
\description{
As the extend of images can changed from one image to another, this function return
the number of crowns included in each image and the vector of the id contained for each id.
}
\examples{

library(sf)

crownsFile <- sf::st_read(
file.path(
system.file(package="managecrownsdata"),
'crowns/Bouamir_crowns.gpkg')
)

path_bbox <- list.files(
file.path(
system.file(package="managecrownsdata"), 'bbox/'),
full.names = TRUE
)

dates <- extr_dates(basename(path_bbox))

check_availableCrowns(path_bbox = path_bbox, crownFile = crownFile, dates = dates)

}
