% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_crownFile.R
\name{check_crownFile}
\alias{check_crownFile}
\title{Check the crowns file compatibility}
\usage{
check_crownFile(crownFile)
}
\arguments{
\item{crownFile}{sf Crown delinetion shapefile.}
}
\value{
NULL. Prints info messages about file compatibility.
}
\description{
Check the crown file compatibility with the \code{canObsR} functions. Your crown
file must have at least the following variables 'id', 'family', 'genus' and 'species'
and it should not have duplicated id.
}
