% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_crownsFile.R
\name{check_crownsFile}
\alias{check_crownsFile}
\title{Check the crowns file compatibility}
\usage{
check_crownsFile(path_crowns)
}
\arguments{
\item{path_crowns}{chr. Path to the crown delinetion shapefile}
}
\value{
Text that give you informations about your file.
Indicates whether your file will be compatible or not  for the other functions of the package.
Pay attention to the line starting with 'X'.
}
\description{
Check the crown file compatibility with the \code{canObsR} functions. Your crown
file must have at least the following variables 'id', 'family', 'genus' and 'species'
and it should not have duplicated id.
}
