% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_files_architecture.R
\name{create_files_architecture}
\alias{create_files_architecture}
\title{Create the recommended files architecture}
\usage{
create_files_architecture(path = NULL, main_folder_name = NULL)
}
\arguments{
\item{path}{chr. The path where to create the main folder}

\item{main_folder_name}{chr. The name of the main folder}
}
\description{
The function just create the recommended files architecture.
}
