% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_labelingFile.R
\name{create_labelingFile}
\alias{create_labelingFile}
\title{Create xlsx labeling file}
\usage{
create_labelingFile(
  crownsFile,
  site = NULL,
  dates = NULL,
  save_xlsx = FALSE,
  directory = NULL
)
}
\arguments{
\item{crownsFile}{\code{sf object}. The crowns delineation file. It should include 'id',
'species', 'genus' and 'family' columns.}

\item{site}{chr. site name.}

\item{dates}{chr. vector of dates (format should be 'YYYY_MM_DD', p.e '20220925').}

\item{save_xlsx}{logical. If TRUE, it will save the table as xlsx file. Indicate the path as the directory parameters}

\item{directory}{The path where to save the xlsx file.}
}
\description{
A function to create the xlsx file to do the labeling
}
