% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_labelingFile.R
\name{create_labelingFile}
\alias{create_labelingFile}
\title{Create xlsx labeling file}
\usage{
create_labelingFile(path_crowns, site, dates, out_dir_path = NULL)
}
\arguments{
\item{path_crowns}{chr. Path to the crown delinetion shapefile}

\item{site}{chr. site name (p.e "Bouamir").}

\item{dates}{chr. vector of dates (format should be 'YYYY_MM_DD', p.e c('2022_09_25','2022_10_10').}

\item{out_dir_path}{chr. The path to the directory used to store the images. By defaut it is NULL,
the data will not be saved but will be return as tibble. If it is not NULL, an xlsx file will be saved.}
}
\value{
A tibble with the variable site, id, family, genus, species, n, obs, update, date, phenophase and comments.
where n, obs, update, phenophase and comments and comments will be NULL. This tibble can be used in the \code{shiny_labels()} applications to be filled.
}
\description{
A function to create the xlsx file where to encode the phenophase.
}
