% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_labelingFile.R
\name{create_labelingFile}
\alias{create_labelingFile}
\title{Create xlsx labeling file}
\usage{
create_labelingFile(
  path_crowns,
  site = NULL,
  dates = NULL,
  save_xlsx = FALSE,
  directory = NULL
)
}
\arguments{
\item{path_crowns}{chr. Path to the crown file}

\item{site}{chr. site name.}

\item{dates}{chr. vector of dates (format should be 'YYYYMMDD', p.e '20220925').}

\item{save_xlsx}{logical. If TRUE, it will save the table as xlsx file. Indicate the path as the directory parameters}

\item{directory}{The path where to save the xlsx file.}
}
\description{
A function to create the xlsx file to do the labeling
}
