% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_labelingFile.R
\name{create_labelingFile}
\alias{create_labelingFile}
\title{Create xlsx labeling file}
\usage{
create_labelingFile(
  crownsFile,
  site = NULL,
  dates = NULL,
  save_xlsx = FALSE,
  directory = NULL
)
}
\arguments{
\item{crownsFile}{\code{sf object}. The crowns delineation file. It should include 'id',
'species', 'genus' and 'family' columns.}

\item{site}{chr. site name.}

\item{dates}{chr. vector of dates (format should be 'YYYY_MM_DD', p.e '20220925').}

\item{save_xlsx}{logical. If TRUE, it will save the table as xlsx file. Indicate the path as the directory parameters}

\item{directory}{The path where to save the xlsx file.}
}
\description{
A function to create the xlsx file to do the labeling
}
\examples{

library(sf)
library(dplyr)
library(stringr)

rgb_paths <- list.files(
file.path(
system.file(package="managecrownsdata"), 'rgb/'),
full.names = TRUE
)
path_crownsFile <- file.path(
system.file(package="managecrownsdata"),
'crowns/Bouamir_crowns.gpkg'
)
crownsFile <- sf::read_sf(path_crownsFile)

site = 'Bouamir'
dates <- paste(
str_sub(stringr::str_split(basename(rgb_paths), '_', simplify = TRUE)[,2],1,4),
str_sub(stringr::str_split(basename(rgb_paths), '_', simplify = TRUE)[,2],5,6),
str_sub(stringr::str_split(basename(rgb_paths), '_', simplify = TRUE)[,2],7,8),
sep = ''
)

check_crownsFile(crownsFile = crownsFile)

crownsFile <- crownsFile \%>\% rename(
geometry = geom
)

check_crownsFile(crownsFile = crownsFile)

create_labelingFile(
crownsFile = crownsFile,
site = site,
dates = dates,
save_xlsx = FALSE
)

}
