% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_labelingFile.R
\name{create_labelingFile}
\alias{create_labelingFile}
\title{Create xlsx labeling file}
\usage{
create_labelingFile(
  crownFile,
  site = NULL,
  date = NULL,
  save_xlsx = FALSE,
  directory = NULL
)
}
\arguments{
\item{crownFile}{A \code{sf} object with the crowns delineation.}

\item{site}{chr. name of the site, p.e 'Mbalmayo'.}

\item{date}{chr. vector of dates (format should be 'yyyy_mm_dd', p.e '2022_09_25').}

\item{save_xlsx}{log. If TRUE, it will save the table as xlsx file. Indicate the path as the directory parameters}

\item{directory}{The path where to save the xlsx file.}
}
\description{
A function to create the xlsx file to do the labeling
}
\examples{
\dontrun{
}

}
