% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_bboxImages.R
\name{extract_bboxImages}
\alias{extract_bboxImages}
\title{Extract and export the non NA area of images.}
\usage{
extract_bboxImages(
  path_in,
  crs = NULL,
  date = NULL,
  directory = NULL,
  filename = NULL
)
}
\arguments{
\item{path_in}{vector with the full paths to the RGB rasters images.}

\item{crs}{crs. Object of class 'crs', could be get from st_crs(..). If NULL,
it will use and transform all the data into the crs of the first RGB image.}

\item{date}{chr. Vector with dates (format should be '\%Y_\%m_\%d', p.e
'2022_09_25'). The order of the dates should match with the order of the
dates of the image in the path_in}

\item{directory}{chr. The path to the directory used to export the vectors.
One file will be export per date. If NULL, the vector will not be exported.}

\item{filename}{logical. The name of the files.}
}
\description{
This function extract the area of images covered by non NA values and return
an \code{sf} object with one polygon per image.
}
