% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_crownsImages.R
\name{extract_crownsImages}
\alias{extract_crownsImages}
\title{Extract crowns images}
\usage{
extract_crownsImages(
  path_images,
  crownFile,
  path_bbox,
  path_out,
  site = NULL,
  date = NULL,
  crs = NULL,
  parallel = FALSE,
  N_cores = NULL,
  update = FALSE,
  specific_quality = FALSE,
  width = 250,
  height = 300
)
}
\arguments{
\item{path_images}{A \code{sf} object for the crowns with an 'id' variable.}

\item{crownFile}{a list with the full paths to the RGB rasters.}

\item{path_bbox}{The path to the non NA Bbox return by the function \code{extract_bboxImages()}}

\item{path_out}{chr. The path to the directory use to stored the images. The
function will create the folder, It doesn't need to exists.}

\item{site}{chr. name of the site, p.e 'Mbalmayo'.}

\item{date}{chr. Vector with dates (format should be '\%Y_\%m_\%d', p.e
'2022_09_25'). The order of the dates should match with the order of the
dates of the image in the path_images}

\item{crs}{crs. Object of class 'crs', could be get from st_crs(..). If NULL,
it will use and transform all the data into the crs of the first RGB image.}

\item{parallel}{xx}

\item{N_cores}{xx}

\item{update}{xx}

\item{specific_quality}{Logical, if TRUE the quality of image will be
determine by the height and width arguments.}

\item{width}{if specific_quality = TRUE, the width of the device}

\item{height}{if specific_quality = TRUE, the height of the device}
}
\description{
The function extracts and save .jpeg images for each crown at
each date.
}
\details{
The extract_crownsImages() create one folder per id and save the
images. The folder names are 'crown_\emph{the id}\emph{\emph{the specie name}' for exemple
'crown_5_Lophira alata'. The the images names are 'crown}\emph{the id}\emph{\emph{the specie
name}}\emph{the date}.jpeg' for exemple 'crown_5_Lophira alata_2022-11-08.jpeg'.
The function upload square image with neighbouring tree and the title is add
at the top, image size is 720*825 pixels. When specific_quality is TRUE, the
image size can be changed by specifying height and width parameters.
}
