% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rgbValues.R
\name{extract_rgbValues}
\alias{extract_rgbValues}
\title{Extract RGB metrics}
\usage{
extract_rgbValues(
  crownsFile,
  path_images,
  site = NULL,
  dates = NULL,
  crs = NULL,
  fun = "all",
  infos = FALSE
)
}
\arguments{
\item{crownsFile}{A \code{sf} object with the crowns delineation.}

\item{path_images}{a list with the full paths to the RGB rasters.}

\item{site}{chr. name of the site, p.e 'Mbalmayo'.}

\item{crs}{crs. Object of class 'crs', could be get from st_crs(..). If NULL, it will
use and transform all the data into the crs of the first RGB image.}

\item{fun}{chr. Specify the function used in the 'fun' parameter of the \code{exactextractr::exact_extract} function to extract
RGB values, it could be 'mean', 'var' or 'all'.}

\item{infos}{logical. Specify whether or not to return details of the extraction.
When TRUE, specify the crowns which has not been extracting per date, because they were out of the image.}

\item{date}{chr. vector of dates (format should be 'yyyy_mm_dd', p.e '2022_09_25').
The order of the dates should match with the order of the path_images !}
}
\description{
The function extracts RGB metrics (red, green, blue, sumrgb,
rcc, gcc, bcc, gndvi, gli) for each crown at each date. It first compute the metrics
for the whole mosaics and then extracts the value at the crown scale using the \code{exactextractr::exact_extract}
function. The mean and / or the variance can be extracted (see 'fun' parameter).
}
