% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rgbValues.R
\name{extract_rgbValues}
\alias{extract_rgbValues}
\title{Extract RGB metrics}
\usage{
extract_rgbValues(
  path_images,
  path_crowns,
  out_dir_path,
  ncor = 1,
  sites = NULL,
  dates = NULL,
  tempdir_custom = NULL
)
}
\arguments{
\item{path_images}{list with the full paths to the RGB rasters.}

\item{path_crowns}{chr. Path to the crown delinetion shapefile}

\item{out_dir_path}{chr. The path to the directory use to stored the images}

\item{ncor}{Number of cores use in the parallelisation proccess.}

\item{sites}{chr. name of the site, p.e 'Mbalmayo'.}

\item{dates}{chr. vector of dates (format should be 'YYYY_MM_DD', p.e '2022_09_25').
The order of the dates should match with the order of the path_images !}

\item{tempdir_custom}{chr. Path where to store temporary files}
}
\value{
A tibble with the variable site, id, date, family, genus, species, type, metric, band, value, plot_name and code_sp.
}
\description{
The function extracts RGB metrics (red, green, blue, sumrgb,
rcc, gcc, bcc, gndvi, gli) for each crown at each date. It extracts the value
at the crown scale using the \code{exactextractr::exact_extract}
function. The mean and / or the variance can be extracted (see 'fun' parameter).
}
\examples{

\dontrun{

imgs = list.files('my-path-to-images', full.names = T)
path_crowns = "my-path-to-crowns-shapefile"
out_dir_path = "output-directory"

rgb_data <- extract_rgbValues (
  path_images = imgs,
  path_crowns = path_crowns,
  out_dir_path = out_dir_path,
  ncor = 10,
  sites = NULL,
  dates = NULL
)
}

}
