% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rgbValues.R
\name{extract_rgbValues}
\alias{extract_rgbValues}
\title{Extract RGB metrics}
\usage{
extract_rgbValues(
  path_in,
  path_crowns,
  out_dir_path,
  ncor = 1,
  sites = NULL,
  dates = NULL
)
}
\arguments{
\item{path_crowns}{chr. Path to the crown file}

\item{out_dir_path}{chr. The path to the directory use to stored the result.}

\item{ncor}{num. Number of cores}

\item{sites}{chr. name of the site, p.e 'Mbalmayo'.}

\item{dates}{chr. vector of dates (format should be 'yyyy_mm_dd', p.e '2022_09_25').
The order of the dates should match with the order of the path_images !}

\item{path_images}{a list with the full paths to the RGB rasters.}
}
\description{
The function extracts RGB metrics (red, green, blue, sumrgb,
rcc, gcc, bcc, gndvi, gli) for each crown at each date. It first compute the metrics
for the whole mosaics and then extracts the value at the crown scale using the \code{exactextractr::exact_extract}
function. The mean and / or the variance can be extracted (see 'fun' parameter).
}
