% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_Labels.R
\name{heatmap_Labels}
\alias{heatmap_Labels}
\title{Plot the heatmap of Labels}
\usage{
heatmap_Labels(
  longLabels,
  Species = NULL,
  Genus = NULL,
  Family = NULL,
  title = NULL,
  na.rm = FALSE
)
}
\arguments{
\item{longLabels}{A \code{tibble} or \code{dataframe} of long data format for labels}

\item{Species}{chr}

\item{Genus}{chr}

\item{Family}{chr}

\item{title}{chr}

\item{na.rm}{logical Remove NA when TRUE}
}
\value{
return a ggplot
}
\description{
A fct function
}
