% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap_Labels.R
\name{heatmap_Labels}
\alias{heatmap_Labels}
\title{Plot the heatmap of Labels}
\usage{
heatmap_Labels(
  longLabels,
  Species = NULL,
  Genus = NULL,
  Family = NULL,
  title = NULL,
  simplify = FALSE,
  repro = FALSE
)
}
\arguments{
\item{longLabels}{\code{tibble} or \code{dataframe} of long data format for labels}

\item{Species}{chr. Specifying the species you want to filter}

\item{Genus}{chr. Specifying the genus you want to filter}

\item{Family}{chr. Specifying the family(ies) you want to filter}

\item{title}{chr. The title of the plot}

\item{simplify}{logical. When TRUE, the plot will use simplified labels instead of raw labels.}

\item{repro}{logical. When TRUE, the flowers and fruits observations will be add to the plot.}
}
\value{
return a ggplot
}
\description{
A fct function
}
\examples{

library(readxl)

raw_labels <- read_excel(
file.path(system.file(package="managecrownsdata"),
'xlsx/labeling_file_Bouamir.xlsx')
)

pivotLabels <- pivot_Labels(raw_labels)

heatmap_Labels(pivotLabels,
Specie = NULL,
Genus = 'Desbordesia',
Family = NULL,
title = NULL)

pivotLabels_simplify <- pivot_Labels(raw_labels, simplify_labels = TRUE)

heatmap_Labels(pivotLabels_simplify,
Specie = NULL,
Genus = 'Desbordesia',
Family = NULL,
title = NULL,
simplify = TRUE)

}
