% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap_Labels.R
\name{heatmap_Labels}
\alias{heatmap_Labels}
\title{Plot the heatmap of Labels}
\usage{
heatmap_Labels(
  Labels,
  Species = NULL,
  Genus = NULL,
  Family = NULL,
  title = NULL,
  simplify = FALSE,
  repro = FALSE
)
}
\arguments{
\item{Labels}{tbl_df.}

\item{Species}{character. Specifying the species you want to filter}

\item{Genus}{character. Specifying the genus you want to filter}

\item{Family}{character. Specifying the family(ies) you want to filter}

\item{title}{character. The title of the plot}

\item{simplify}{logical. When TRUE, the plot will use simplified labels instead of raw labels.}

\item{repro}{logical. When TRUE, the flowers and fruits observations will be add to the plot.}
}
\value{
ggplot
}
\description{
A fct function
}
