% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_values.R
\name{merge_values}
\alias{merge_values}
\title{Merge labels and rgb values}
\usage{
merge_values(Labels, rgb_data, out_dir_path = NULL, file_type = ".RData")
}
\arguments{
\item{Labels}{tbl_df. Labels data}

\item{rgb_data}{tbl_df. rgb metrics values.}

\item{out_dir_path}{character. Directory where the outputs are saved.}

\item{file_type}{character. By default it is '.RData' but can be '.csv' or '.xlsx'}
}
\value{
A tibble with the variable site, id, date, family, genus, species, phenophase, type, metric, band, value, obs, comments, update, Usable_crown.
}
\description{
A function to merge the labels from long format and the rgb metrics
values.
}
\examples{

library(canObsR)

data_labeling = canObsR::data_labeling
rgb_data = canObsR::rgb_data
merge_values(data_labeling,rgb_data)


}
