% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_Labels.R
\name{pivot_Labels}
\alias{pivot_Labels}
\title{Format labels values from wide data to long data}
\usage{
pivot_Labels(labels_path, simplify_labels = FALSE, out_dir_path = NULL)
}
\arguments{
\item{labels_path}{character. Path to the labeling file}

\item{simplify_labels}{logical. Decompose and simplify the labels when TRUE. By defaut it is FALSE.}

\item{out_dir_path}{character. Directory where the outputs are saved.}
}
\value{
tbl_df
}
\description{
A function to format labels data from wide to long format
}
