% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_Labels.R
\name{pivot_Labels}
\alias{pivot_Labels}
\title{Reformat table of labels from wide data to long data}
\usage{
pivot_Labels(wideLabels, simplify_labels = FALSE)
}
\arguments{
\item{wideLabels}{\code{tibble} or \code{dataframe} which contains labeling data directly import from the xlsx file.}

\item{simplify_labels}{\code{logical} Decompose and simplify the labels when TRUE. By defaut it is FALSE.}
}
\value{
\code{tibble}
}
\description{
A function to format labels data from wide to long format
}
\examples{

library(tidyverse)
library(readxl)

raw_labels <- read_excel(file.path(system.file(package="managecrownsdata"), 'xlsx/labeling_file_Bouamir.xlsx'))

pivot_Labels(raw_labels)

pivotLabels <- pivot_Labels(raw_labels, simplify_labels = TRUE)

}
