% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signal.R
\name{plot_signal}
\alias{plot_signal}
\title{Plot the spectral signal over time}
\usage{
plot_signal(
  data,
  Species = NULL,
  Genus = NULL,
  Family = NULL,
  Type = NULL,
  Metric = "mean",
  Band = NULL,
  facet_by = "band",
  slcted_id = NULL,
  show_Labels = FALSE,
  title = NULL
)
}
\arguments{
\item{data}{A \code{tibble} or \code{dataframe} of long data format for labels}

\item{Species}{chr Filtering the data at the species level}

\item{Genus}{chr Filtering the data at the genus level}

\item{Family}{chr Filtering the data at the family level}

\item{Type}{chr Filtering by the type}

\item{Metric}{chr Filtering by the metric. By defaut it is the mean.}

\item{Band}{chr Filtering by the band.}

\item{facet_by}{chr Facetting by a variable}

\item{slcted_id}{chr Highlighting an id}

\item{show_Labels}{description}

\item{title}{chr The title of the plot}
}
\value{
return a ggplot
}
\description{
A fct function
}
