% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signal.R
\name{plot_signal}
\alias{plot_signal}
\title{Plot the spectral signal over time}
\usage{
plot_signal(
  data,
  Species = NULL,
  Genus = NULL,
  Family = NULL,
  Type = NULL,
  Metric = "mean",
  Band = NULL,
  facet_by = "band",
  slcted_id = NULL,
  show_Labels = FALSE,
  title = NULL
)
}
\arguments{
\item{data}{tbl_df. Labels Labels}

\item{Species}{character. Specifying the species you want to filter}

\item{Genus}{character. Specifying the genus you want to filter}

\item{Family}{character. Specifying the family(ies) you want to filter}

\item{Type}{character. Specifying the type(s) you want to filter}

\item{Metric}{character. Specifying the metric(s) you want to filter. By defaut 'mean'}

\item{Band}{character. Specifying the bande(s) you want to filter}

\item{facet_by}{character. Facetting by a variable. By defaut 'band'}

\item{slcted_id}{numeric. Highlighting an id}

\item{show_Labels}{logical. When TRUE, it shows the phenophase labels on the plot.}

\item{title}{character. The title of the plot}
}
\value{
return a ggplot
}
\description{
A fct function
}
