% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signal.R
\name{plot_signal}
\alias{plot_signal}
\title{Plot the spectral signal over time}
\usage{
plot_signal(
  data,
  Species = NULL,
  Genus = NULL,
  Family = NULL,
  Type = NULL,
  Metric = "mean",
  Band = NULL,
  facet_by = "band",
  slcted_id = NULL,
  show_Labels = FALSE,
  title = NULL
)
}
\arguments{
\item{data}{\code{tibble} or \code{dataframe} of long data format for labels}

\item{Species}{chr. Specifying the species you want to filter}

\item{Genus}{chr. Specifying the genus you want to filter}

\item{Family}{chr. Specifying the family(ies) you want to filter}

\item{Type}{chr. Specifying the type(s) you want to filter}

\item{Metric}{chr. Specifying the metric(s) you want to filter. By defaut 'mean'}

\item{Band}{chr. Specifying the bande(s) you want to filter}

\item{facet_by}{chr. Facetting by a variable. By defaut 'band'}

\item{slcted_id}{num. Highlighting an id}

\item{show_Labels}{logical. When TRUE, it shows the phenophase labels on the plot.}

\item{title}{chr. The title of the plot}
}
\value{
return a ggplot
}
\description{
A fct function
}
\examples{

library(readxl)

data('data')

plot_signal(data = merge_data, Genus = 'Desbordesia')

plot_signal(data = merge_data, Genus = 'Desbordesia', slcted_id = 676, Band = c('gli', 'gndvi'))

plot_signal(data = merge_data, Genus = 'Desbordesia', Band = c('gli'), slcted_id = 676, show_Labels = TRUE)


}
