% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{rgb_data}
\alias{rgb_data}
\title{RGB indices data}
\format{
A data.frame with 225888 observations on the following 12 variables:
\itemize{
\item \code{site}: site
\item \code{id}: Crown id
\item \code{date}: date of the rgb survey
\item \code{family}: family
\item \code{genus}: genus
\item \code{species}: species
\item \code{type}: type
\item \code{metric}: metric
\item \code{band}: band
\item \code{value}: value
\item \code{plot_name}: Plot name
\item \code{code_sp}: Species code
}
}
\usage{
data("rgb_data")
}
\description{
RGB indices data
}
\examples{
data(rgb_data)
}
\keyword{datasets}
\keyword{internal}
