% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_labels.R
\name{shiny_labels}
\alias{shiny_labels}
\title{Labeling shiny app}
\usage{
shiny_labels(data_labeling, newFile, imgFolder)
}
\arguments{
\item{data_labeling}{chr. Path to crown images folder.}

\item{newFile}{chr. Path to save the new labeling file.}
}
\description{
Shiny app to do the labeling based on crowns images.
}
\examples{

library(canObsR)

data(data_labeling)
data_labeling
# shiny_labels(data_labeling = data_labeling) # Run it

}
