% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_labels.R, R/shiny_visualisation.R
\name{shiny_labels}
\alias{shiny_labels}
\title{Labeling shiny app}
\usage{
shiny_labels(data_labeling)

shiny_labels(data_labeling)
}
\arguments{
\item{data_labeling}{tibble. Labeling file resulting from the \code{\link[canObsR]{create_labelingFile}}}
}
\description{
Shiny app to do the labeling based on crowns images.

Shiny app to do the visualise the crowns data (labels and rgb indices).
}
\examples{

library(canObsR)

data(data_labeling)
data_labeling
# shiny_labels(data_labeling = data_labeling) # Run it


library(canObsR)

# shiny_visualisation() # Run it

}
