% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_cores.R
\name{Time_SIFT_in_r}
\alias{Time_SIFT_in_r}
\title{Check the crown file compatibility}
\usage{
Time_SIFT_in_r(
  pathDIR,
  out_dir_ortho,
  out_dir_DEM = NULL,
  out_dir_project = NULL,
  data_type = "RGB",
  resol_ref = 0.05,
  crs = "EPSG::32622",
  site_name = "",
  calibrate_col = TRUE,
  sun_sensor = FALSE,
  group_by_flight = FALSE,
  downscale_factor_alignement = 1L,
  downscale_factor_depth_map = 2L
)
}
\arguments{
\item{crownFile}{A sf object}
}
\value{
Text that give you information about your file.
Indicates whether your file will be compatible or not  for the other functions of the package.
Pay attention to the line starting with '-- ERROR --'.
}
\description{
Check the crown file compatibility
}
